Use slower (but more cross-platform) method of invoking shell commands
Different unixen have different maximum lengths for command line so when
`cp * /some/place` was expanded by the shell it would exceed that limit.

From now we iterate over files in dirs and invokes a separate instance of
the command for each file - this is slower but more reliable.

This commit also makes the script more verbose in its output so the user
can track progress more easily